Project files related to all previous commits.
TODO: Fix the gradient in the comments cell. Look horrible atm.
TODO: Add a Tweetbot like tooltip bar on click on a Comment cell.
TODO: Add a option button in the right top corner in the CommentVC in
order to navigate back to the story and more.
GTG: Test on iPads.